Update Band-Band.py
i am just reading through the code, it is very promising. one comment is that maybe you can combine the draw function of each type with the preview functions in a way that you only have one function for drawing a module and the first time the codes run it just receives some default values to draw one of each outside the grid. think about it.
💯